+++ mukurobot/__init__.py
@@ -5,7 +5,7 @@ from discord import Intents
 import requests
 import sys
 
-__version__ = '0.2.0'
+__version__ = '0.2.1'
 
 dotenv.load_dotenv()